split meshes.ts into mesh-list.ts and mesh-loader.ts
