+++ src/cannons/bullet-collision.ts
@@ -10,7 +10,7 @@ import { PhysicsResultsDef } from "../physics/nonintersection.js";
 import { AuthorityDef } from "../net/components.js";
 import { BulletDef } from "./bullet.js";
 import { DeletedDef } from "../ecs/delete.js";
-import { AllMeshesDef } from "../meshes/meshes.js";
+import { AllMeshesDef } from "../meshes/mesh-list.js";
 import { AudioDef } from "../audio/audio.js";
 import { PositionDef, RotationDef } from "../physics/transform.js";
 import { assert, NumberTuple } from "../utils/util.js";
