+++ src/cloth/cloth.ts
@@ -5,7 +5,7 @@ import { PositionDef, RotationDef } from "../physics/transform.js";
 import { SyncDef, AuthorityDef, Me, MeDef } from "../net/components.js";
 import { Serializer, Deserializer } from "../utils/serialize.js";
 import { FinishedDef } from "../ecs/em-helpers.js";
-import { AllMeshes, AllMeshesDef } from "../meshes/meshes.js";
+import { AllMeshes, AllMeshesDef } from "../meshes/mesh-list.js";
 import { SpringType, SpringGridDef, ForceDef } from "./spring.js";
 import {
   Mesh,
