+++ src/cloth/game-cloth.ts
@@ -36,7 +36,7 @@ import { tempVec3 } from "../matrix/temp-pool.js";
 import { EASE_INQUAD } from "../utils/util-ease.js";
 import { assert } from "../utils/util.js";
 import { drawLine } from "../utils/utils-game.js";
-import { AllMeshesDef } from "../meshes/meshes.js";
+import { AllMeshesDef } from "../meshes/mesh-list.js";
 import { ClothConstructDef, ClothLocalDef } from "./cloth.js";
 import { GlobalCursor3dDef } from "../gui/cursor.js";
 import { ENEMY_SHIP_COLOR } from "../hyperspace/uv-enemy-ship.js";
