+++ src/debug/xp-cube.ts
@@ -22,7 +22,7 @@ import {
   canvasTexturePtr,
 } from "../render/pipelines/std-scene.js";
 import { uintToVec3unorm } from "../utils/utils-3d.js";
-import { AllMeshesDef } from "../meshes/meshes.js";
+import { AllMeshesDef } from "../meshes/mesh-list.js";
 import { GlobalCursor3dDef } from "../gui/cursor.js";
 import { createGhost } from "./ghost.js";
 import { PointLightDef } from "../render/lights.js";
