+++ src/grass/game-grass.ts
@@ -2,12 +2,8 @@ import { CameraDef, CameraFollowDef } from "../camera/camera.js";
 import { ColorDef } from "../color/color-ecs.js";
 import { ENDESGA16 } from "../color/palettes.js";
 import { EM } from "../ecs/entity-manager.js";
-import {
-  BallMesh,
-  GizmoMesh,
-  UnitCubeMesh,
-  defineMeshSetResource,
-} from "../meshes/meshes.js";
+import { BallMesh, GizmoMesh, UnitCubeMesh } from "../meshes/mesh-list.js";
+import { defineMeshSetResource } from "../meshes/mesh-loader.js";
 import { ControllableDef } from "../input/controllable.js";
 import { createGhost, GhostDef } from "../debug/ghost.js";
 import { LocalHsPlayerDef, HsPlayerDef } from "../hyperspace/hs-player.js";
