+++ src/grass/grass.ts
@@ -1,6 +1,5 @@
 import { ColorDef } from "../color/color-ecs.js";
 import { EM, EntityW } from "../ecs/entity-manager.js";
-import { gameMeshFromMesh } from "../meshes/meshes.js";
 import { grassPoolPtr, RenderDataGrassDef } from "./std-grass.js";
 import { jitter, align } from "../utils/math.js";
 import { PositionDef } from "../physics/transform.js";
