+++ src/gui/path-editor.ts
@@ -1,6 +1,7 @@
 import { ColorDef } from "../color/color-ecs.js";
 import { EM, EntityW } from "../ecs/entity-manager.js";
-import { AllMeshesDef, GameMesh, gameMeshFromMesh } from "../meshes/meshes.js";
+import { AllMeshesDef } from "../meshes/mesh-list.js";
+import { GameMesh, gameMeshFromMesh } from "../meshes/mesh-loader.js";
 import { gameplaySystems } from "../debug/ghost.js";
 import { vec2, vec3, vec4, quat, mat4, V } from "../matrix/sprig-matrix.js";
 import {
