+++ src/hyperspace/hs-player.ts
@@ -27,7 +27,7 @@ import { Ray } from "../physics/broadphase.js";
 import { copyAABB, createAABB } from "../physics/aabb.js";
 import { tempVec3 } from "../matrix/temp-pool.js";
 import { cloneMesh, scaleMesh3 } from "../meshes/mesh.js";
-import { AllMeshesDef } from "../meshes/meshes.js";
+import { AllMeshesDef } from "../meshes/mesh-list.js";
 import { LinearVelocityDef } from "../motion/velocity.js";
 import { MotionSmoothingDef } from "../render/motion-smoothing.js";
 import { ModelerDef } from "../meshes/modeler.js";
