+++ src/hyperspace/uv-enemy-ship.ts
@@ -14,7 +14,8 @@ import {
 import { ColliderDef } from "../physics/collider.js";
 import { AuthorityDef, MeDef } from "../net/components.js";
 import { aabbCenter } from "../physics/aabb.js";
-import { AllMeshes, AllMeshesDef, GameMesh } from "../meshes/meshes.js";
+import { AllMeshes, AllMeshesDef } from "../meshes/mesh-list.js";
+import { GameMesh } from "../meshes/mesh-loader.js";
 import { AngularVelocityDef, LinearVelocityDef } from "../motion/velocity.js";
 import { MotionSmoothingDef } from "../render/motion-smoothing.js";
 import {
