+++ src/ld53/game-ld53.ts
@@ -2,7 +2,7 @@ import { CameraDef, CameraFollowDef } from "../camera/camera.js";
 import { ColorDef } from "../color/color-ecs.js";
 import { ENDESGA16 } from "../color/palettes.js";
 import { EM, EntityW } from "../ecs/entity-manager.js";
-import { AllMeshesDef, GizmoMesh } from "../meshes/meshes.js";
+import { AllMeshesDef, GizmoMesh } from "../meshes/mesh-list.js";
 import { ControllableDef } from "../input/controllable.js";
 import { createGhost, GhostDef } from "../debug/ghost.js";
 import {
