+++ src/meshes/half-edge.ts
@@ -1,5 +1,5 @@
 import { DBG_ASSERT } from "../flags.js";
-import { BLACK } from "./meshes.js";
+import { BLACK } from "./mesh-list.js";
 import { vec2, vec3, vec4, quat, mat4, V } from "../matrix/sprig-matrix.js";
 import { hexAvg } from "../hex/hex.js";
 import { RawMesh } from "./mesh.js";
