+++ src/physics/game-rebound.ts
@@ -15,7 +15,8 @@ import {
 import { RendererDef } from "../render/renderer-ecs.js";
 import { assert } from "../utils/util.js";
 import { TimeDef } from "../time/time.js";
-import { AllMeshesDef, GameMesh } from "../meshes/meshes.js";
+import { AllMeshesDef } from "../meshes/mesh-list.js";
+import { GameMesh } from "../meshes/mesh-loader.js";
 // import { ENEMY_SHIP_COLOR } from "./enemy-ship.js";
 // import { ClothConstructDef, ClothLocalDef } from "./cloth.js";
 import { GlobalCursor3dDef } from "../gui/cursor.js";
