+++ src/physics/phys-debug.ts
@@ -1,6 +1,6 @@
 import { ColliderDef } from "./collider.js";
 import { EM } from "../ecs/entity-manager.js";
-import { AllMeshesDef } from "../meshes/meshes.js";
+import { AllMeshesDef } from "../meshes/mesh-list.js";
 import { ColorDef } from "../color/color-ecs.js";
 import { InputsDef } from "../input/inputs.js";
 import { mathMap } from "../utils/math.js";
@@ -26,7 +26,7 @@ import {
 } from "./transform.js";
 import { vec3, V } from "../matrix/sprig-matrix.js";
 import { Phase } from "../ecs/sys-phase.js";
-import { CUBE_MESH } from "../meshes/primatives";
+import { CUBE_MESH } from "../meshes/primatives.js";
 
 // TODO(@darzu): re-enable all this! it requires line drawing again
